+++ Sources/VCore/Deprecations.swift
@@ -4,3 +4,46 @@
 //
 //  Created by Vakhtang Kontridze on 10/9/21.
 //
+
+// MARK: - Architectural Pattern Helpers
+import SwiftUI
+
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
+extension ProgressViewParameters {
+    @available(*, deprecated, message: "Use init without default value for `isInteractionDisabled`")
+    public init(
+        isInteractionDisabled: Bool = false,
+        scalingFactor: CGFloat?
+    ) {
+        self.init(
+            scalingFactor: scalingFactor,
+            color: nil,
+            isInteractionDisabled: isInteractionDisabled
+        )
+    }
+    
+    @available(*, deprecated, message: "Use init without default value for `isInteractionDisabled`")
+    public init(
+        isInteractionDisabled: Bool = false,
+        color: Color?
+    ) {
+        self.init(
+            scalingFactor: nil,
+            color: color,
+            isInteractionDisabled: isInteractionDisabled
+        )
+    }
+    
+    @available(*, deprecated, message: "Use init without default value for `isInteractionDisabled`")
+    public init(
+        isInteractionDisabled: Bool = false,
+        scalingFactor: CGFloat?,
+        color: Color?
+    ) {
+        self.init(
+            scalingFactor: scalingFactor,
+            color: color,
+            isInteractionDisabled: isInteractionDisabled
+        )
+    }
+}