+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Views
 - `SecurableTextField` now takes `Text` as placeholder
 - `ProgressViewParameters` now causes `View` to ignore interaction, instead of disabling it
 
+Architectural Pattern Helpers
+
+- `isInteractionDisabled` no longer has a default value in `ProgressViewParameters`
+
 ### [4.0.1(64)](https://github.com/VakhoKontridze/VCore/releases/tag/4.0.1) — *2022 09 15*
 
 Services and Managers